app/requirements.txt
huggingface_hub
llama-cpp-python
streamlit
openfabric-pysdk
sentence-transformers
pinecone
accelerate
fastapi
pydantic
uvicorn
watchdog
python-dotenv